module for output
