+++ lib/runner.rb
@@ -1,7 +1,11 @@
 require_relative "event_reporter"
 require "pry"
+require_relative 'output'
 
 class Runner
+
+  include Output
+
   attr_reader :reporter, :queue
 
   def initialize
@@ -10,19 +14,8 @@ class Runner
     evaluate_first_command(input_splitter)
   end
 
-  def intro
-    puts "Hello please type 'load' and the file you would like to load"
-  end
-
   def input_splitter
     STDIN.gets.chomp.split(" ")
-  end
-
-  def restart_message
-    puts "That is not an executable command.
-    Please enter a valid csv file after
-    load or we will load the full_event_attendees.csv
-    if you type load and then hit enter."
   end
 
   def evaluate_first_command(input)
@@ -56,8 +49,7 @@ class Runner
   end
 
   def try_again_next_command
-    puts "That is not a valid command. Enter a valid command
-    or type 'help' to see a list of valid commands"
+    restart_message_with_help
     evaluate_next_command(input_splitter)
   end
